pyproject.toml
[project]
name = "printke"
version = "1.0.0"
description = "Kenya's Premier Online Card Printing Service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-login>=0.6",
    "flask-wtf>=1.2",
    "flask-cors>=4.0",
    "pillow>=10.0",
    "img2pdf>=0.5",
    "python-dotenv>=1.0",
    "gunicorn>=21.0",
    "requests>=2.31",
    "qrcode>=7.4",
    "celery>=5.3",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "email-validator>=2.1",
    "python-magic>=0.4",
    "werkzeug>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "flake8>=6.1",
]
